scripts/pyproject.toml
[project]
name = "dwh-scripts"
version = "1.0.0"
description = "Scripts utilitaires pour le Data Warehouse Azure"
requires-python = ">=3.10"
dependencies = [
    "pyodbc>=5.0.0",
    "faker>=20.0.0",
    "python-dotenv>=1.0.0",
    "sh>=2.0.0",
    "azure-eventhub>=5.11.0",
    "azure-storage-blob>=12.19.0",
]
